requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.4.0
PyQt6==6.2.0
PyQt6-Qt6==6.2.0
PyQt6-sip==13.1.0
